tests/wpt/web-platform-tests/app-history/app-history-entry/sameDocument-after-navigate-restore.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  await i.contentWindow.appHistory.navigate("#foo");
  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  assert_true(i.contentWindow.appHistory.entries()[0].sameDocument);

  i.contentWindow.appHistory.navigate("/common/blank.html?bar");
  await new Promise(resolve => i.onload = resolve);
  assert_equals(i.contentWindow.appHistory.entries().length, 3);
  assert_false(i.contentWindow.appHistory.entries()[0].sameDocument);
  assert_false(i.contentWindow.appHistory.entries()[1].sameDocument);

  i.contentWindow.appHistory.back();
  await new Promise(resolve => i.onload = resolve);
  assert_equals(i.contentWindow.appHistory.entries().length, 3);
  assert_equals(i.contentWindow.appHistory.current.index, 1);
  assert_true(i.contentWindow.appHistory.entries()[0].sameDocument);
  assert_false(i.contentWindow.appHistory.entries()[2].sameDocument);
}, "entry.sameDocument is properly restored after cross-document back");
</script>
